requirements.txt
blobfile>=1.0.5
torch
tqdm
numpy
scipy
requests
tqdm
ftfy
attrs
Pillow
filelock
regex

